Initial commit plus vertices
Vertices where included as a Newtype wrapper of Int. Vertices in
in a graph start at 0 and can go up to a predefined maximum.
Due to the plans to use single integers for the internal
representation of an edge, a Vertex can go up to a maximum of
sqrt(maxBound :: Int), so on 32-bit systems, vertices can only
go up to about 46'000, while on 64-bit system they can go up
to 3 billion.